mv _strtok and _strobrk funcs to parser file
